backend/requirements.txt
alembic==1.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
httpie==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
